+++ modules/cactus-web/src/Table/Table.tsx
@@ -1,7 +1,7 @@
-import { ColorStyle, Shape, TextStyle } from '@repay/cactus-theme'
+import { CactusTheme, ColorStyle, Shape, TextStyle } from '@repay/cactus-theme'
 import PropTypes from 'prop-types'
 import React, { createContext, useContext, useLayoutEffect } from 'react'
-import styled, { css, FlattenSimpleInterpolation } from 'styled-components'
+import styled, { css, FlattenSimpleInterpolation, ThemeProps } from 'styled-components'
 import { width, WidthProps } from 'styled-system'
 
 import { useMergedRefs } from '../helpers/react'
@@ -20,6 +20,7 @@ interface TableContextProps {
   headers?: React.ReactNode[]
   cellIndex: number
   variant: TableVariant
+  dividers?: boolean
 }
 
 interface TableProps
@@ -28,6 +29,7 @@ interface TableProps
   cardBreakpoint?: Size
   variant?: TableVariant
   as?: React.ElementType
+  dividers?: boolean
 }
 
 interface TableHeaderProps
@@ -36,6 +38,7 @@ interface TableHeaderProps
     HTMLTableSectionElement
   > {
   variant?: TableVariant
+  dividers?: boolean
 }
 
 export interface TableCellProps
@@ -67,6 +70,7 @@ const DEFAULT_CONTEXT: TableContextProps = {
   inHeader: false,
   cellIndex: 0,
   variant: 'table',
+  dividers: false,
 }
 
 const TableContext = createContext<TableContextProps>(DEFAULT_CONTEXT)
@@ -127,6 +131,9 @@ export const Table = React.forwardRef<HTMLTableElement, TableProps>(
     } else if (cardBreakpoint && size <= SIZES[cardBreakpoint]) {
       context.variant = 'card'
     }
+    if (props.dividers) {
+      context.dividers = props.dividers
+    }
     props.variant = context.variant
     return (
       <TableContext.Provider value={context}>
@@ -185,7 +192,7 @@ export const TableHeader = React.forwardRef<HTMLTableSectionElement, TableHeader
     const context = useContext<TableContextProps>(TableContext)
     return (
       <TableContext.Provider value={{ ...context, inHeader: true, cellType: 'th', cellIndex: 0 }}>
-        <StyledHeader {...props} variant={context.variant} ref={ref}>
+        <StyledHeader {...props} variant={context.variant} ref={ref} dividers={context.dividers}>
           <tr>{children}</tr>
         </StyledHeader>
       </TableContext.Provider>
@@ -294,7 +301,7 @@ const StyledCell = styled.td(
               ${media(p.theme, 'extraLarge')} {
                 min-width: 160px;
               }
-            `}
+            `};
     `,
     card: css`
       && {
@@ -333,17 +340,34 @@ const StyledHeader = styled.thead<TableHeaderProps>`
   &&&&& td {
     text-transform: uppercase;
     border: ${(p): string => border(p.theme, 'base')};
+    border-right: ${(p): string => (p.dividers ? border(p.theme, 'mediumContrast') : '')};
     ${(p): ColorStyle => p.theme.colorStyles.base};
   }
   ${headerVariants}
 `
 
-const table = css`
+const getCTABorder = (p: ThemeProps<CactusTheme>, focus?: boolean): ReturnType<typeof css> => {
+  return css`
+    th,
+    td {
+      border-left-color: ${p.theme.colors.callToAction};
+      border-top-color: ${p.theme.colors.callToAction};
+      border-bottom-color: ${p.theme.colors.callToAction};
+      background-color: ${focus && p.theme.colors.transparentCTA};
+    }
+    td:last-child {
+      border-color: ${p.theme.colors.callToAction};
+    }
+  `
+}
+
+const table = css<TableProps>`
   display: table;
-  ${(p): FlattenSimpleInterpolation | TextStyle => textStyle(p.theme, 'small')};
+  ${(p): FlattenSimpleInterpolation | TextStyle => textStyle(p.theme, 'body')};
   border-spacing: 0;
   td,
   th {
+    border-right: ${(p): string => (p.dividers ? border(p.theme, 'lightContrast') : '')};
     background-color: ${(p): string => p.theme.colors.white};
     border-top: ${(p): string => border(p.theme, 'transparent')};
     border-bottom: ${(p): string => border(p.theme, 'transparent')};
@@ -359,20 +383,17 @@ const table = css`
     th {
       background-color: ${(p): string => p.theme.colors.lightContrast};
     }
-  }
-  &&& tr:hover {
-    th,
-    td {
-      border-color: ${(p): string => p.theme.colors.callToAction};
+    td:not(:last-child) {
+      border-right: ${(p): string => (p.dividers ? border(p.theme, 'white') : '')};
     }
+  }
+
+  &&& tr:hover {
+    ${getCTABorder}
   }
   &&& tr:focus {
     outline: 0;
-    td,
-    th {
-      background-color: ${(p): string => p.theme.colors.transparentCTA};
-      border-color: ${(p): string => p.theme.colors.callToAction};
-    }
+    ${(p) => getCTABorder(p, true)}
   }
   // first row
   & > tr:first-of-type,