+++ modules/cactus-web/src/Toggle/Toggle.story.tsx
@@ -9,28 +9,18 @@ export default {
   component: Toggle,
 } as Meta
 
-const initialState = { value: false }
-type State = Readonly<typeof initialState>
-
-class ToggleManager extends React.Component {
-  public readonly state: State = initialState
-  private handleToggle = (): void => {
-    this.setState({
-      value: !this.state.value,
-    })
-  }
-  public render(): React.ReactElement {
-    return (
-      <Toggle
-        value={this.state.value}
-        onClick={this.handleToggle}
-        disabled={boolean('disabled', false)}
-        {...this.props}
-      />
-    )
-  }
+const ToggleManager = (props: any) => {
+  const [checked, setChecked] = React.useState<boolean>(false)
+  const onChange = React.useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      console.log(`onChange '${e.target.name}': ${e.target.checked}`)
+      setChecked(e.target.checked)
+    },
+    [setChecked]
+  )
+  return <Toggle {...props} checked={checked} onChange={onChange} />
 }
 
 export const BasicUsage = (): React.ReactElement => {
-  return <ToggleManager />
+  return <ToggleManager disabled={boolean('Disabled', false)} />
 }