+++ modules/cactus-web/src/Toggle/Toggle.test.tsx
@@ -1,15 +1,20 @@
 import { generateTheme } from '@repay/cactus-theme'
-import { fireEvent, render } from '@testing-library/react'
+import { render } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
+import pick from 'lodash/pick'
 import * as React from 'react'
 
 import { StyleProvider } from '../StyleProvider/StyleProvider'
 import Toggle from './Toggle'
 
+// Use this to avoid PropTypes error regarding readonly field.
+const noop = () => undefined
+
 describe('component: Toggle', (): void => {
   test('should render a toggle', (): void => {
     const toggle = render(
       <StyleProvider>
-        <Toggle value={false} />
+        <Toggle checked={false} onChange={noop} />
       </StyleProvider>
     )
 
@@ -19,17 +24,17 @@ describe('component: Toggle', (): void => {
   test('should render a disabled toggle', (): void => {
     const toggle = render(
       <StyleProvider>
-        <Toggle value={false} disabled={true} />
+        <Toggle checked={false} disabled={true} />
       </StyleProvider>
     )
 
     expect(toggle.asFragment()).toMatchSnapshot()
   })
 
-  test('should initialize value to true', (): void => {
+  test('should initialize checked to true', (): void => {
     const toggle = render(
       <StyleProvider>
-        <Toggle value={true} />
+        <Toggle checked={true} onChange={noop} />
       </StyleProvider>
     )
 
@@ -39,35 +44,37 @@ describe('component: Toggle', (): void => {
   test('should support margin space props', (): void => {
     const toggle = render(
       <StyleProvider>
-        <Toggle value={false} marginBottom={4} />
+        <Toggle checked={false} marginBottom={4} onChange={noop} />
       </StyleProvider>
     )
 
     expect(toggle.asFragment()).toMatchSnapshot()
   })
 
-  test('should trigger onClick event', (): void => {
-    const onClick = jest.fn()
-    const { getByTestId } = render(
+  test('should trigger onChange event', (): void => {
+    const box: any = {}
+    const onChange = jest.fn((e) => Object.assign(box, pick(e.target, ['id', 'checked'])))
+    const { container } = render(
       <StyleProvider>
-        <Toggle value={false} onClick={onClick} data-testid="will-click" />
+        <Toggle checked={false} onChange={onChange} id="will-change" />
       </StyleProvider>
     )
 
-    fireEvent.click(getByTestId('will-click'))
-    expect(onClick).toHaveBeenCalled()
+    userEvent.click(container.querySelector('#will-change') as Element)
+    expect(onChange).toHaveBeenCalled()
+    expect(box).toEqual({ id: 'will-change', checked: true })
   })
 
-  test('should not trigger onClick event', (): void => {
-    const onClick = jest.fn()
+  test('should not trigger onChange event', (): void => {
+    const onChange = jest.fn()
     const { getByTestId } = render(
       <StyleProvider>
-        <Toggle value={false} onClick={onClick} disabled={true} data-testid="will-not-click" />
+        <Toggle disabled onChange={onChange} data-testid="will-not-change" />
       </StyleProvider>
     )
 
-    fireEvent.click(getByTestId('will-not-click'))
-    expect(onClick).not.toHaveBeenCalled()
+    userEvent.click(getByTestId('will-not-change'))
+    expect(onChange).not.toHaveBeenCalled()
   })
 
   describe('with theme customization', (): void => {