+++ modules/cactus-icons/CHANGELOG.md
@@ -1,3 +1,27 @@
+# v2.1.3 (Mon Nov 30 2020)
+
+#### 🔩 Dependency Updates
+
+- Upgrade `@repay/cactus-theme` to `v2.0.0`
+
+#### Authors: 1
+
+- Mike Soltow ([@mikesoltow](https://github.com/mikesoltow))
+
+---
+
+# v2.1.3-beta.0 (Thu Nov 19 2020)
+
+#### 🔩 Dependency Updates
+
+- Upgrade `@repay/cactus-theme` to `v2.0.0-beta-0`
+
+#### Authors: 1
+
+- Mike Soltow ([@mikesoltow](https://github.com/mikesoltow))
+
+---
+
 # v2.1.2 (Wed Nov 11 2020)
 
 #### 🐛 Bug Fix